Don't allow values that doesn't make sense.

+++ src/ops.rs
@@ -2,19 +2,19 @@ use std::ops::Div;
 use crate::{time::*, frequency::*, magnitude::*};
 
 /// The inverse of Time is Frequency.
-impl Div<Time> for f64 {
+impl Div<Time> for i32 {
     type Output = Frequency;
 
     fn div(self, rhs: Time) -> Self::Output {
-        Frequency::si(self / rhs.val())
+        Frequency::si(self as f64 / rhs.val())
     }
 }
 
 /// The inverse of Frequency is Time.
-impl Div<Frequency> for f64 {
+impl Div<Frequency> for i32 {
     type Output = Time;
 
     fn div(self, rhs: Frequency) -> Self::Output {
-        Time::si(self / rhs.val())
+        Time::si(self as f64 / rhs.val())
     }
 }
